[storage] simplify wake_and_wait_pruner
