+++ storage/aptosdb/src/pruner/ledger_pruner_manager.rs
@@ -43,6 +43,12 @@ pub(crate) struct LedgerPrunerManager {
 }
 
 impl PrunerManager for LedgerPrunerManager {
+    type Pruner = LedgerPruner;
+
+    fn pruner(&self) -> &Self::Pruner {
+        &self.pruner
+    }
+
     fn is_pruner_enabled(&self) -> bool {
         self.pruner_enabled
     }
@@ -89,40 +95,6 @@ impl PrunerManager for LedgerPrunerManager {
         self.pruner_worker
             .as_ref()
             .set_target_db_version(latest_version.saturating_sub(self.prune_window));
-    }
-
-    #[cfg(test)]
-    fn wake_and_wait_pruner(&self, latest_version: Version) -> anyhow::Result<()> {
-        use std::{
-            thread::sleep,
-            time::{Duration, Instant},
-        };
-
-        *self.latest_version.lock() = latest_version;
-
-        if latest_version
-            >= *self.last_version_sent_to_pruner.as_ref().lock() + self.pruning_batch_size as u64
-        {
-            self.set_pruner_target_db_version(latest_version);
-            *self.last_version_sent_to_pruner.as_ref().lock() = latest_version;
-        }
-
-        if self.pruner_enabled && latest_version > self.prune_window {
-            let min_readable_ledger_version = latest_version - self.prune_window;
-
-            // Assuming no big pruning chunks will be issued by a test.
-            const TIMEOUT: Duration = Duration::from_secs(10);
-            let end = Instant::now() + TIMEOUT;
-
-            while Instant::now() < end {
-                if self.get_min_readable_version() >= min_readable_ledger_version {
-                    return Ok(());
-                }
-                sleep(Duration::from_millis(1));
-            }
-            anyhow::bail!("Timeout waiting for pruner worker.");
-        }
-        Ok(())
     }
 }
